Revert "Merge branch 'android-3.18' of https://android.googlesource.com/kernel/common into lineage-18.1"
This reverts commit 68f07e267baeed2cb94e5f664403424567a03fe6.

+++ fs/file.c
@@ -62,7 +62,7 @@ static void free_fdtable_rcu(struct rcu_head *rcu)
  */
 static void copy_fdtable(struct fdtable *nfdt, struct fdtable *ofdt)
 {
-	size_t cpy, set;
+	unsigned int cpy, set;
 
 	BUG_ON(nfdt->max_fds < ofdt->max_fds);
